simplify iso2 lookup
